- **IndexBuffer 유연성 개선**:   - `Uint16Array` 지원 추가를 통해 인덱스 버퍼 데이터 구조 확장.   - `GPUIndexFormat` 반환 로직 업데이트로 데이터 타입에 따라 동적 포맷 처리.   - 코드 내 데이터 타입 관련 주석 및 선언 정리.
브랜치: V4.4.0-Alpha-rnd

+++ src/display/terrain/core/geometry/TerrainGeometry.ts
@@ -106,7 +106,7 @@ class TerrainGeometry extends Geometry {
 
         const indexBuffer = new IndexBuffer(
             redGPUContext,
-            new Uint32Array(indexData),
+            new Uint16Array(indexData),
             undefined,
             `IndexBuffer_TerrainGeometry`
         );
